refactor: start separating OAuth in its own files
