+++ .yarn/sdks/typescript/package.json
@@ -1,6 +1,6 @@
 {
   "name": "typescript",
-  "version": "5.1.3-sdk",
+  "version": "5.1.6-sdk",
   "main": "./lib/typescript.js",
   "type": "commonjs"
 }
