+++ src/services/generic-service.ts
@@ -1,6 +1,6 @@
-import { Cache } from '../util'
 import { ItemTypes, ItemReturns, PaginatedResponse, PaginatedItemType, PaginatedItemOptions,
   PaginatedType, Resolvable, PaginatedInstance, PaginatedRequestParams } from '../types'
+import { Cache } from '../oauth'
 import YouTube,
 { Video, Channel, Playlist, Comment, VideoAbuseReportReason, Subscription, VideoCategory, Language, Region, ChannelSection, Caption } from '..'
