+++ src/services/search-service.ts
@@ -1,4 +1,4 @@
-import { Cache } from '../util'
+import { Cache } from '../oauth'
 import YouTube, { Video, Channel, Playlist, InternalSearchOptions, SearchType, PaginatedResponse, SearchFilters, VideoCategory } from '..'
 
 /**
